Moved data-signing from connector to app and simplified function api a bit
